remove gfx2 namespace
