+++ source/ncengine/graphics2/frontend/subsystem/animation/CMakeLists.txt
@@ -1,8 +1,8 @@
 target_sources(${NC_ENGINE_LIB}
     PRIVATE
         BoneCache.cpp
+        Rig.cpp
         SkeletalAnimationCalculator.cpp
         SkeletalAnimationSubsystem.cpp
         SkeletalAnimationStorage.cpp
-        SkeletalAnimationTypes.cpp
 )
